requirements.txt
cmake
pybind11
protobuf
scikit-build
setuptools_scm
build
grpcio==1.64.1
grpcio-tools==1.62.2
